refactor: remove snapshot profile shim
