+++ packages/react/runtime/src/snapshot/lifecycle/destroy.ts
@@ -7,7 +7,7 @@ import { __root } from '../../root.js';
 import { delayedEvents } from './event/delayEvents.js';
 import { delayedLifecycleEvents } from './event/delayLifecycleEvents.js';
 import { globalCommitTaskMap } from './patch/commit.js';
-import { profileEnd, profileStart } from '../debug/profile.js';
+import { profileEnd, profileStart } from '../../shared/profile.js';
 
 function destroyBackground(): void {
   if (typeof __PROFILE__ !== 'undefined' && __PROFILE__) {
