+++ packages/react/runtime/src/snapshot/lifecycle/reload.ts
@@ -13,8 +13,8 @@ import { destroyBackground } from './destroy.js';
 import { increaseReloadVersion } from './pass.js';
 import { renderMainThread } from './render.js';
 import { __root, setRoot } from '../../root.js';
+import { profileEnd, profileStart } from '../../shared/profile.js';
 import { isEmptyObject } from '../../utils.js';
-import { profileEnd, profileStart } from '../debug/profile.js';
 import { LifecycleConstant } from '../lifecycle/constant.js';
 import { __pendingListUpdates } from '../list/pendingListUpdates.js';
 import { hydrate } from '../renderToOpcodes/hydrate.js';
