+++ packages/react/runtime/src/snapshot/lynx/env.ts
@@ -2,7 +2,7 @@
 // Licensed under the Apache License Version 2.0 that can be found in the
 // LICENSE file in the root directory of this source tree.
 import type { DataProcessorDefinition, InitData, InitDataRaw } from '../../lynx-api.js';
-import { profileEnd, profileStart } from '../debug/profile.js';
+import { profileEnd, profileStart } from '../../shared/profile.js';
 
 export function setupLynxEnv(): void {
   if (!__LEPUS__) {
