+++ packages/react/runtime/src/snapshot/lynx/tt.ts
@@ -6,9 +6,9 @@ import { process, render } from 'preact';
 import { PerformanceTimingFlags, PipelineOrigins, beginPipeline, markTiming } from './performance.js';
 import { runWithForce } from './runWithForce.js';
 import { __root } from '../../root.js';
+import { profileEnd, profileStart } from '../../shared/profile.js';
 import { CHILDREN } from '../../shared/render-constants.js';
 import { printSnapshotInstanceToString } from '../debug/printSnapshot.js';
-import { profileEnd, profileStart } from '../debug/profile.js';
 import { getSnapshotVNodeSource } from '../debug/vnodeSource.js';
 import { LifecycleConstant, NativeUpdateDataType } from '../lifecycle/constant.js';
 import type { FirstScreenData } from '../lifecycle/constant.js';
