+++ packages/react/runtime/src/snapshot/snapshot/backgroundSnapshot.ts
@@ -20,8 +20,8 @@ import { hydrationMap } from './snapshotInstanceHydrationMap.js';
 import { transformSpread } from './spread.js';
 import type { SerializedSnapshotInstance } from './types.js';
 import { traverseSnapshotInstance } from './utils.js';
+import { profileEnd, profileStart } from '../../shared/profile.js';
 import { isDirectOrDeepEqual } from '../../utils.js';
-import { profileEnd, profileStart } from '../debug/profile.js';
 import { clearSnapshotVNodeSource, getSnapshotVNodeSource, moveSnapshotVNodeSource } from '../debug/vnodeSource.js';
 import { prepareGestureForCommit } from '../gesture/processGestureBagkround.js';
 import type { GestureKind } from '../gesture/types.js';
